moving withTypes to procedures
