+++ src/server/procedures.ts
@@ -1,7 +1,6 @@
 import { z } from "zod";
 import { withProtectedSession } from "~/server/auth/withSession";
 import type { RequestEventLoader } from "~/server/types";
-import { withTypedParams } from "~/server/withTypes";
 import { paths } from "~/utils/paths";
 
 type ProcedureBuilderResult<R extends RequestEventLoader> = {
@@ -56,6 +55,42 @@ export const procedureBuilder =
       use: (middle) => procedureBuilderInner((e) => middle(e)),
     };
   };
+
+export const withTypedParams = <
+  P extends z.ZodRawShape = z.ZodRawShape,
+  R extends RequestEventLoader = RequestEventLoader
+>(
+  schema: z.ZodObject<P>
+) => {
+  return (event: R) => {
+    const query = schema?.safeParse(event.params);
+
+    if (!query.success) {
+      throw event.redirect(302, "/404");
+    }
+
+    return { ...event, typedParams: query.data };
+  };
+};
+
+export const withTypedQuery = <
+  Q extends z.ZodRawShape = z.ZodRawShape,
+  R extends RequestEventLoader = RequestEventLoader
+>(
+  schema: z.ZodObject<Q>
+) => {
+  return (event: R) => {
+    const rawQuery = Object.fromEntries(event.url.searchParams.entries());
+
+    const query = schema?.safeParse(rawQuery);
+
+    if (!query.success) {
+      throw event.redirect(302, "/404");
+    }
+
+    return { ...event, query: query.data };
+  };
+};
 
 export const protectedProcedure = procedureBuilder().use(
   withProtectedSession()
